Delete job by ID
Deleting a Job By id, Here used a method in controller which gets the id should delete from the website, in the job Service interface there s method called deleteJobById and it is implemented in the JobServiceimpl class. It has used iterator to iterate through the job list and find corresponding id and delete the job. and then returns a true value if the job deleted and returns false if it is not