Use entities with JPA
Use @Entity annotation to make a object an entity, apart from that we can use @GeneratedValue(strategy = GenerationType.IDENTITY)   this to auto increment the value of the id(primary key), After that we also need to make a default constructor for the class when making an entity. its because entities are objects cuz its an  essential.